Use pages directly
Without dynamic imports
